requirements.txt
asgiref==3.7.2
Django==4.2.8
django-jazzmin==2.6.0
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
